+++ hook.php
@@ -9,7 +9,7 @@ function plugin_version_projecthelper()
 {
     return [
         'name' => 'Project Helper',
-        'version' => '1.2.1',
+        'version' => '1.3.0',
         'author' => 'Joao-SouzaDev',
         'license' => 'AGPLv3+',
         'homepage' => 'https://github.com/Joao-SouzaDev/projecthelper',
@@ -49,7 +49,6 @@ function plugin_init_projecthelper()
 
     // Hook para desinstalação.
     $PLUGIN_HOOKS['uninstall']['projecthelper'] = [Install::class, 'uninstall'];
-    $PLUGIN_HOOKS['update']['projecthelper'] = [Install::class, 'update'];
 
     // Hook para replicar followups quando adicionados a um ticket
     $PLUGIN_HOOKS['item_add']['projecthelper'] = [